# Optimize `Range#bsearch` for beginless/endless ranges