@nobu 2023-09-14
Without changes to `left` and `right`, is this adding assignments of `mid` to `beg` and `end` at `while` loops respectively?