@kyanagi 2023-09-14
> Without changes to `left` and `right`, is this adding assignments of `mid` to `beg` and `end` at `while` loops respectively?

Exactly. In this pull request, I avoided overwriting `beg` and `end`, but it seems that overwriting them would be more efficient.
I'll update the PR.

```diff
diff --git a/range.c b/range.c
index 62e957e622..b18a25ea35 100644
--- a/range.c
+++ b/range.c
@@ -758,6 +758,7 @@ range_bsearch(VALUE range)
                 return bsearch_integer_range(beg, mid, 0);
             }
             diff = rb_funcall(diff, '*', 1, LONG2FIX(2));
+            beg = mid;
         }
     }
     else if (NIL_P(beg) && is_integer_p(end)) {
@@ -770,6 +771,7 @@ range_bsearch(VALUE range)
                 return bsearch_integer_range(mid, end, 0);
             }
             diff = rb_funcall(diff, '*', 1, LONG2FIX(2));
+            end = mid;
         }
     }
     else {
```

```
% ~/tmp/r/bin/ruby ~/bm.rb
Warming up --------------------------------------
               10**1     1.387k i/100ms
               10**2   810.000  i/100ms
               10**3   566.000  i/100ms
               10**4   418.000  i/100ms
               10**5   343.000  i/100ms
Calculating -------------------------------------
               10**1     13.812k (± 0.8%) i/s -     69.350k in   5.021492s
               10**2      8.119k (± 0.4%) i/s -     41.310k in   5.088215s
               10**3      5.672k (± 0.6%) i/s -     28.866k in   5.089336s
               10**4      4.182k (± 0.5%) i/s -     21.318k in   5.098243s
               10**5      3.442k (± 0.6%) i/s -     17.493k in   5.082207s
```